UPSTREAM: firmware: smccc: Add support for erratum discovery API
It is not always possible for the OS to determine if a CPU is affected
by a particular erratum. For example, it may depend on an integration
choice the chip designer made, or whether firmware has enabled some
particular feature.

Add support for the SMCCC 'Errata Management Firmware Interface' that
lets the OS query firmware for this information.

Link: https://developer.arm.com/documentation/den0100/1-0/?lang=en

Change-Id: I0e3a2b3d180dcce6dbb651aba170293204e8718a
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>